DarkUI: Touchups
App info buttons:
Change the default style give it a darker look

The blueman group have left the building..
No idea why google left the On-body detection icon tint to blue but to
me it was way outta place.
Added a default tint "accent_device_default_light" in the suggestion
tile layout to keep it at least looking tidy and most likely help
themers in the long run.
Thanks to @daveyannihilation for the tips in the past and leading me
in the proper direction for icon tint.